move facial recognition service into person service
