+++ server/src/domain/facial-recognition/index.ts
@@ -1,2 +1,1 @@
 export * from './face.repository';
-export * from './facial-recognition.services';
